added clang-format, formated stuff, added Union; NOT STABLE, WILL NOT WORK

+++ VCore/Utils/HashUtils.h
@@ -1,11 +1,10 @@
 #pragma once
 /*
  * MIT LICENSE
- * Copyright (c) 2019 Vladyslav Joss 
+ * Copyright (c) 2019 Vladyslav Joss
  */
-#include <random>
-
 #include <cstdint>
+#include <random>
 
 #if INTPTR_MAX == INT64_MAX
 #define ECSCORE_x64
@@ -16,9 +15,9 @@
 #endif
 
 #pragma warning(push)
-#pragma warning( disable : 26495 ) 
-#pragma warning( disable : 26451 ) 
-#pragma warning( disable : 28812 ) 
+#pragma warning(disable : 26495)
+#pragma warning(disable : 26451)
+#pragma warning(disable : 28812)
 #include <VCore/Dependencies/MurmurHash3.h>
 #pragma warning(pop)
 
@@ -26,22 +25,18 @@ namespace vex
 {
 	namespace util
 	{
-		static constexpr int gPrimeNumbers[] = {
-			3, 7, 11, 17, 23, 29, 37, 47, 59, 71, 89, 107, 131,
-			163, 197, 239, 293, 353, 431, 521, 631, 761, 919,
-			1103, 1327, 1597, 1931, 2333, 2801, 3371, 4049, 4861,
-			5839, 7013, 8419, 10103, 12143, 14591, 17321, 21269,
-			25253, 31393, 39769, 49157, 62851, 90523, 108631, 130363, 156437,
-			187751, 225307, 270371, 324449, 389357, 467237, 560689, 672827, 807403, 946037,
-			1395263, 1572869, 2009191, 2411033, 2893249, 3471899, 4166287, 4999559, 5999471, 7199369
-		};
+		static constexpr int gPrimeNumbers[] = {3, 7, 11, 17, 23, 29, 37, 47, 59, 71, 89, 107, 131, 163, 197, 239, 293,
+			353, 431, 521, 631, 761, 919, 1103, 1327, 1597, 1931, 2333, 2801, 3371, 4049, 4861, 5839, 7013, 8419, 10103,
+			12143, 14591, 17321, 21269, 25253, 31393, 39769, 49157, 62851, 90523, 108631, 130363, 156437, 187751,
+			225307, 270371, 324449, 389357, 467237, 560689, 672827, 807403, 946037, 1395263, 1572869, 2009191, 2411033,
+			2893249, 3471899, 4166287, 4999559, 5999471, 7199369};
 
 		static constexpr int gPrimeSize = sizeof(gPrimeNumbers) / sizeof(int);
 
 		inline constexpr int FindUpperBound(const int* a, int n, int x)
 		{
-			if (x > a[n-1])
-				return a[n-1];
+			if (x > a[n - 1])
+				return a[n - 1];
 
 			int iter = 0;
 			int count = n;
@@ -73,17 +68,13 @@ namespace vex
 			return gPrimeNumbers[gPrimeSize - 1];
 		}
 
-		inline constexpr int ClosestPrimeSearch(int value)
-		{
-			return FindUpperBound(gPrimeNumbers, gPrimeSize, value);
-		}
+		inline constexpr int ClosestPrimeSearch(int value) { return FindUpperBound(gPrimeNumbers, gPrimeSize, value); }
 
 		inline int RandomRange(int fromInc, int toExc)
 		{
 			static std::random_device rd;
 			static std::mt19937 mt(rd());
-			std::uniform_int_distribution<int> dist(
-				fromInc,
+			std::uniform_int_distribution<int> dist(fromInc,
 				toExc - 1); // exclusive, as it is more common use case
 			return dist(mt);
 		}
@@ -94,8 +85,7 @@ namespace vex
 		static inline int fnv1a(std::string const& text)
 		{
 			std::size_t hash = fnv_offset_basis;
-			for (std::string::const_iterator it = text.begin(), end = text.end();
-				it != end; ++it)
+			for (std::string::const_iterator it = text.begin(), end = text.end(); it != end; ++it)
 			{
 				hash ^= *it;
 				hash *= fnv_prime;
@@ -103,17 +93,15 @@ namespace vex
 
 			return (int)hash;
 		}
-		static inline int Hash(char* c, int sz)
-		{
-			return (int)murmur::MurmurHash3_x86_32(c, sz);
-		}
+
+		static inline int Hash(char* c, int sz) { return (int)murmur::MurmurHash3_x86_32(c, sz); }
 
 		struct SHash
 		{
 			static inline int Hash(const std::string& str)
 			{
 #ifdef ECSCORE_x64
-				return murmur::MurmurHash3_x86_32(str.data(), str.size());
+				return (int)murmur::MurmurHash3_x86_32(str.data(), (int)str.size());
 #else
 				return fnv1a(str);
 #endif
@@ -121,17 +109,11 @@ namespace vex
 		};
 		struct SHash_STD
 		{
-			static inline int Hash(const std::string& str)
-			{
-				return (int)std::hash<std::string>{}(str);
-			}
+			static inline int Hash(const std::string& str) { return (int)std::hash<std::string>{}(str); }
 		};
 		struct SHash_FNV1a
 		{
-			static inline int Hash(const std::string& str)
-			{
-				return fnv1a(str);
-			}
+			static inline int Hash(const std::string& str) { return fnv1a(str); }
 		};
 		struct SHash_MURMUR
 		{
@@ -140,5 +122,5 @@ namespace vex
 				return (int)murmur::MurmurHash3_x86_32(str.data(), (int)str.size());
 			}
 		};
-	}
-}
+	} // namespace util
+} // namespace vex
